% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabetise_wlp_file.R
\name{alphabetise_wlp_file}
\alias{alphabetise_wlp_file}
\title{(Re-)alphabetise the Warlpiri dictionary file}
\usage{
alphabetise_wlp_file(wlp_lexicon, output_file)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
(Re-)alphabetise the Warlpiri dictionary file
}
