% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wlp_lexicon.R
\name{read_wlp_lexicon}
\alias{read_wlp_lexicon}
\title{Read in and optionally pre-process the Warlpiri lexicon file}
\usage{
read_wlp_lexicon(lexicon_path, preprocess = TRUE)
}
\arguments{
\item{lexicon_path}{location of the wlp-lexicon_master.txt file.}

\item{preprocess}{extract out the left-most backslash code of each line into a \code{code1} column? Default is \code{TRUE}}
}
\description{
Read in and optionally pre-process the Warlpiri lexicon file
}
