% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_all_codes_defined.R
\name{test_all_codes_defined}
\alias{test_all_codes_defined}
\title{Test whether all left-most codes are defined (i.e. no stray codes from typos)}
\usage{
test_all_codes_defined(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test whether all left-most codes are defined (i.e. no stray codes from typos)
}
