% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_blacklist_chars.R
\name{test_blacklist_chars}
\alias{test_blacklist_chars}
\title{Test that no blacklisted characters appear in the lexicon}
\usage{
test_blacklist_chars(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that no blacklisted characters appear in the lexicon
}
