% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_block_attributes.R
\name{test_block_attributes}
\alias{test_block_attributes}
\title{Test that listed attributes for entry blocks are well-formatted}
\usage{
test_block_attributes(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that listed attributes for entry blocks are well-formatted
}
