% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_brackets_balanced.R
\name{test_brackets_balanced}
\alias{test_brackets_balanced}
\title{Test that all brackets and parentheses are well-balanced}
\usage{
test_brackets_balanced(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that all brackets and parentheses are well-balanced
}
