% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code1_ordered.R
\name{test_code1_ordered}
\alias{test_code1_ordered}
\title{Test whether left-most codes appear in expected order within the entries}
\usage{
test_code1_ordered(wlp_lexicon, return_all = FALSE,
  grammar_file = system.file("structures/wlp_skeleton-simple.ne", package
  = "yinarlingi"))
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}

\item{return_all}{return the all from data frame or just entries with errors (default is \code{FALSE})}

\item{grammar_file}{a Nearley grammar file (default is `wlp_skeleton-simple.ne`)}
}
\description{
Test whether left-most codes appear in expected order within the entries
}
