% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_entries_unique.R
\name{test_entries_unique}
\alias{test_entries_unique}
\title{Test that all main- and sub-entry forms are unique}
\usage{
test_entries_unique(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that all main- and sub-entry forms are unique
}
