% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_homophone_sequences.R
\name{test_homophone_sequences}
\alias{test_homophone_sequences}
\title{Test that all homophones numbers well-sequenced across the homophonous entry forms}
\usage{
test_homophone_sequences(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that all homophones numbers well-sequenced across the homophonous entry forms
}
