% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_images_ok.R
\name{test_images_ok}
\alias{test_images_ok}
\title{Test that all listed images in Dropbox/WarlpiriDictionaryFiles2017/wlp-dictionary-illustrations}
\usage{
test_images_ok(wlp_lexicon, file_list)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}

\item{file_list}{output of rclone file listing}
}
\description{
Test that all listed images in Dropbox/WarlpiriDictionaryFiles2017/wlp-dictionary-illustrations
}
