% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inline_codes_ok.R
\name{test_inline_codes_ok}
\alias{test_inline_codes_ok}
\title{Test whether all expected codes (and only these) codes appear on a given line}
\usage{
test_inline_codes_ok(wlp_df)
}
\arguments{
\item{wlp_df}{a Warlpiri lexicon data frame}
}
\description{
Test whether all expected codes (and only these) codes appear on a given line
}
