% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_latin_ok.R
\name{test_latin_ok}
\alias{test_latin_ok}
\title{Test that latin/scientific name(s) lines are well-formed}
\usage{
test_latin_ok(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that latin/scientific name(s) lines are well-formed
}
