% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_linestarts_ok.R
\name{test_linestarts_ok}
\alias{test_linestarts_ok}
\title{Test that all non-empty lines start with either a backslash or \code{@}}
\usage{
test_linestarts_ok(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that all non-empty lines start with either a backslash or \code{@}
}
