% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pos_strings.R
\name{test_pos_strings}
\alias{test_pos_strings}
\title{Test that parts of speech values within main- and sub-entry lines are well-formatted}
\usage{
test_pos_strings(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that parts of speech values within main- and sub-entry lines are well-formatted
}
