% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pvl_fuzzy_homed.R
\name{test_pvl_fuzzy_homed}
\alias{test_pvl_fuzzy_homed}
\title{Test that cross-references to preverbs have a fuzzily-matching parent}
\usage{
test_pvl_fuzzy_homed(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
For a given preverb, e.g. 'jurlkuly-' we search for whether various approximate forms, e.g.
'jurlkuly', 'jurlkulyku', 'jurlkuly-ku', 'jurlkuly(pa)', or 'jurlkulypa' can be found as a
main- or sub-entry (e.g. \code{\\me jurlkuly(pa)})
}
