% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_quotes_balanced.R
\name{test_quotes_balanced}
\alias{test_quotes_balanced}
\title{Test that double quotes are well-balanced}
\usage{
test_quotes_balanced(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that double quotes are well-balanced
}
