% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_reversals_ok.R
\name{test_reversals_ok}
\alias{test_reversals_ok}
\title{Test that reversal mark up in reversal and gloss lines are well-formed}
\usage{
test_reversals_ok(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that reversal mark up in reversal and gloss lines are well-formed
}
