% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_values_nonempty.R
\name{test_values_nonempty}
\alias{test_values_nonempty}
\title{Test whether all non-block codes have some content in them}
\usage{
test_values_nonempty(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test whether all non-block codes have some content in them
}
