% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_xrefs_noselfref.R
\name{test_xrefs_noselfref}
\alias{test_xrefs_noselfref}
\title{Test that cross-references do not contain references to themselves}
\usage{
test_xrefs_noselfref(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
Test that cross-references do not contain references to themselves
}
