% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_wlp_regex.R
\name{use_wlp_regex}
\alias{use_wlp_regex}
\title{Use a pre-defined regular expression for the Warlpiri Dictionary}
\usage{
use_wlp_regex(regex_name = NA)
}
\arguments{
\item{regex_name}{Name of regular expression to get, leave as NA to see the names. \link{yinarlingi}}
}
\description{
Use a pre-defined regular expression for the Warlpiri Dictionary
}
