% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wlp_images.R
\name{view_wlp_images}
\alias{view_wlp_images}
\title{View list of linked images in the Warlpiri dictionary}
\usage{
view_wlp_images(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
View list of linked images in the Warlpiri dictionary
}
