% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wlp_kinrev.R
\name{view_wlp_kinrev}
\alias{view_wlp_kinrev}
\title{View Warlpiri (kin terms)-to-English finder list}
\usage{
view_wlp_kinrev(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
View Warlpiri (kin terms)-to-English finder list
}
