% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wlp_reversals.R
\name{view_wlp_reversals}
\alias{view_wlp_reversals}
\title{View English-to-Warlpiri finder list}
\usage{
view_wlp_reversals(wlp_lexicon)
}
\arguments{
\item{wlp_lexicon}{a Warlpiri lexicon data frame, or path to a Warlpiri dictionary file}
}
\description{
View English-to-Warlpiri finder list
}
\examples{
system.file("extdata/wlp-lexicon_bigger.txt", package = "yinarlingi") \%>\%
   view_wlp_reversals()
}
