% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yinarlingi.R
\docType{package}
\name{yinarlingi}
\alias{yinarlingi}
\alias{yinarlingi-package}
\title{Yinarlingi package}
\description{
Google spreadsheets R API
}
\details{
See the README on
\href{https://cran.r-project.org/package=googlesheets/README.html}{CRAN}
or \href{https://github.com/jennybc/googlesheets#readme}{GitHub}
}
