% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wlp_groups.R
\name{add_wlp_groups}
\alias{add_wlp_groups}
\title{Add grouping columns to a Warlpiri lexicon data frame}
\usage{
add_wlp_groups(wlp_df, groupings)
}
\arguments{
\item{wlp_df}{a Warlpiri lexicon data frame}

\item{groupings}{type of grouping(s) to append to the data frame}
}
\description{
Add grouping columns to a Warlpiri lexicon data frame
}
