% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_xrefs_homed.R
\name{test_xrefs_homed}
\alias{test_xrefs_homed}
\title{Test that no cross-references are orphaned in the Warlpiri dictionary}
\usage{
test_xrefs_homed(wlp_df)
}
\arguments{
\item{wlp_df}{a Warlpiri lexicon data frame}
}
\description{
Test that no cross-references are orphaned in the Warlpiri dictionary
}
