% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wlp_reversals.R
\name{view_wlp_reversals}
\alias{view_wlp_reversals}
\title{View English-to-Warlpiri finder list}
\usage{
view_wlp_reversals(wlp_df)
}
\arguments{
\item{wlp_df}{a Warlpiri lexicon data frame}
}
\description{
View English-to-Warlpiri finder list
}
